refactor: ToolbarActivity & ToolbarFragment 정의
