+++ android/app/src/main/java/poke/rogue/helper/presentation/ability/AbilityFragment.kt
@@ -11,13 +11,13 @@ import poke.rogue.helper.data.datasource.RemoteAbilityDataSource
 import poke.rogue.helper.data.repository.DefaultAbilityRepository
 import poke.rogue.helper.databinding.FragmentAbilityBinding
 import poke.rogue.helper.presentation.ability.detail.AbilityDetailFragment
-import poke.rogue.helper.presentation.base.BindingFragment
+import poke.rogue.helper.presentation.toolbar.ToolbarFragment
 import poke.rogue.helper.presentation.util.repeatOnStarted
 import poke.rogue.helper.presentation.util.view.LinearSpacingItemDecoration
 import poke.rogue.helper.presentation.util.view.dp
 import poke.rogue.helper.remote.ServiceModule
 
-class AbilityFragment : BindingFragment<FragmentAbilityBinding>(R.layout.fragment_ability) {
+class AbilityFragment : ToolbarFragment<FragmentAbilityBinding>(R.layout.fragment_ability) {
     private val viewModel by viewModels<AbilityViewModel> {
         AbilityViewModel.factory(
             DefaultAbilityRepository(
@@ -31,14 +31,14 @@ class AbilityFragment : BindingFragment<FragmentAbilityBinding>(R.layout.fragmen
 
     private val adapter: AbilityAdapter by lazy { AbilityAdapter(viewModel) }
 
-    override fun toolbar(): Toolbar = binding.toolbarAbility
+    override val toolbar: Toolbar?
+        get() = binding.toolbarAbility
 
     override fun onViewCreated(
         view: View,
         savedInstanceState: Bundle?,
     ) {
         super.onViewCreated(view, savedInstanceState)
-
         binding.vm = viewModel
         binding.lifecycleOwner = viewLifecycleOwner
