+++ android/app/src/main/java/poke/rogue/helper/presentation/ability/detail/AbilityDetailFragment.kt
@@ -11,8 +11,8 @@ import poke.rogue.helper.data.datasource.RemoteAbilityDataSource
 import poke.rogue.helper.data.repository.DefaultAbilityRepository
 import poke.rogue.helper.databinding.FragmentAbilityDetailBinding
 import poke.rogue.helper.presentation.ability.model.toUi
-import poke.rogue.helper.presentation.base.BindingFragment
 import poke.rogue.helper.presentation.dex.detail.PokemonDetailFragment
+import poke.rogue.helper.presentation.toolbar.ToolbarFragment
 import poke.rogue.helper.presentation.util.fragment.toast
 import poke.rogue.helper.presentation.util.repeatOnStarted
 import poke.rogue.helper.presentation.util.view.GridSpacingItemDecoration
@@ -20,7 +20,7 @@ import poke.rogue.helper.presentation.util.view.dp
 import poke.rogue.helper.remote.ServiceModule
 
 class AbilityDetailFragment :
-    BindingFragment<FragmentAbilityDetailBinding>(R.layout.fragment_ability_detail) {
+    ToolbarFragment<FragmentAbilityDetailBinding>(R.layout.fragment_ability_detail) {
     private val viewModel by viewModels<AbilityDetailViewModel> {
         AbilityDetailViewModel.factory(
             DefaultAbilityRepository(
@@ -34,7 +34,8 @@ class AbilityDetailFragment :
 
     private val adapter: AbilityDetailAdapter by lazy { AbilityDetailAdapter(viewModel) }
 
-    override fun toolbar(): Toolbar = binding.toolbarAbilityDetail
+    override val toolbar: Toolbar?
+        get() = binding.toolbarAbilityDetail
 
     override fun onViewCreated(
         view: View,
