+++ android/app/src/main/java/poke/rogue/helper/presentation/dex/PokemonListFragment.kt
@@ -11,15 +11,15 @@ import poke.rogue.helper.R
 import poke.rogue.helper.data.datasource.RemotePokemonListDataSource
 import poke.rogue.helper.data.repository.DefaultPokemonListRepository
 import poke.rogue.helper.databinding.FragmentPokemonListBinding
-import poke.rogue.helper.presentation.base.BindingFragment
 import poke.rogue.helper.presentation.dex.detail.PokemonDetailFragment
+import poke.rogue.helper.presentation.toolbar.ToolbarFragment
 import poke.rogue.helper.presentation.util.repeatOnStarted
 import poke.rogue.helper.presentation.util.view.GridSpacingItemDecoration
 import poke.rogue.helper.presentation.util.view.dp
 import poke.rogue.helper.remote.ServiceModule
 
 class PokemonListFragment :
-    BindingFragment<FragmentPokemonListBinding>(R.layout.fragment_pokemon_list) {
+    ToolbarFragment<FragmentPokemonListBinding>(R.layout.fragment_pokemon_list) {
     private val viewModel by viewModels<PokemonListViewModel> {
         PokemonListViewModel.factory(
             pokemonListRepository =
@@ -36,14 +36,14 @@ class PokemonListFragment :
         PokemonAdapter(viewModel)
     }
 
-    override fun toolbar(): Toolbar = binding.toolbarDex
+    override val toolbar: Toolbar?
+        get() = binding.toolbarDex
 
     override fun onViewCreated(
         view: View,
         savedInstanceState: Bundle?,
     ) {
         super.onViewCreated(view, savedInstanceState)
-
         binding.viewModel = viewModel
         binding.lifecycleOwner = viewLifecycleOwner
